+++ lib/savon/options.rb
@@ -268,6 +268,10 @@ module Savon
       @options[:ssl_ca_cert] = cert
     end
 
+    def ssl_ciphers(ciphers)
+      @options[:ssl_ciphers] = ciphers
+    end
+
     # Sets the ca cert path.
     def ssl_ca_cert_path(path)
       @options[:ssl_ca_cert_path] = path